# backend/generated/flask-server/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: "MVP v1 REST API for ACME Inc. to ingest shipping disruption signals,\
    \ enrich relevance and impact, escalate incidents, and generate action proposals\
    \ for manager decisioning."
  title: Shipping Disruption Management API
  version: 1.0.0
servers:
- url: /api/v1
...
paths:
...
  /incidents:
    get:
      operationId: list_incidents
      parameters:
      - explode: true
        in: query
        name: status
        required: false
        schema:
          $ref: "#/components/schemas/IncidentStatus"
        style: form
      - explode: true
        in: query
        name: classification
        required: false
        schema:
          $ref: "#/components/schemas/IncidentClassification"
        style: form
      - explode: true
        in: query
        name: risk_level
        required: false
        schema:
          $ref: "#/components/schemas/RiskLevel"
        style: form
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentListResponse"
          description: Paginated incidents
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: List incidents
      tags:
      - Incidents
      x-openapi-router-controller: openapi_server.controllers.incidents_controller
  /incidents/{incident_id}:
    get:
      operationId: get_incident
      parameters:
      - explode: false
        in: path
        name: incident_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
          description: Incident
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Get incident by id
      tags:
      - Incidents
      x-openapi-router-controller: openapi_server.controllers.incidents_controller
...
  /reference/ports:
    get:
      operationId: list_ports
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceListResponse"
          description: Port reference list
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: List port reference data
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
  /reference/routes:
    get:
      operationId: list_routes
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceListResponse"
          description: Route reference list
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: List route reference data
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
  /reference/shipments:
    get:
      operationId: list_shipments
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentListResponse"
          description: Shipment reference list
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: List shipment reference data
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
    post:
      operationId: create_shipment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
          description: Shipment created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Invalid request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Semantic validation failed
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Create shipment reference record
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
  /reference/shipments/{shipment_id}:
    patch:
      operationId: update_shipment
      parameters:
      - explode: false
        in: path
        name: shipment_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
          description: Shipment updated
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Semantic validation failed
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Update shipment reference record
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
  /reference/skus:
    get:
      operationId: list_skus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuListResponse"
          description: SKU reference list
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: List SKU reference data
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
    post:
      operationId: create_sku
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkuRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sku"
          description: SKU created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Invalid request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Semantic validation failed
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Create SKU reference record
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
  /reference/skus/{sku_id}:
    patch:
      operationId: update_sku
      parameters:
      - explode: false
        in: path
        name: sku_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSkuRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sku"
          description: SKU updated
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Semantic validation failed
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Update SKU reference record
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
  /reference/suppliers:
    get:
      operationId: list_suppliers
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierListResponse"
          description: Supplier reference list
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: List supplier reference data
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
    post:
      operationId: create_supplier
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
          description: Supplier created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Invalid request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Semantic validation failed
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Create supplier reference record
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
  /reference/suppliers/{supplier_id}:
    patch:
      operationId: update_supplier
      parameters:
      - explode: false
        in: path
        name: supplier_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
          description: Supplier updated
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Resource conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Semantic validation failed
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Update supplier reference record
      tags:
      - Reference
      x-openapi-router-controller: openapi_server.controllers.reference_controller
components:
...
  schemas:
    ErrorEnvelope:
      example:
        error:
          code: INVALID_ARGUMENT
          details:
            key: ""
          message: human readable
      properties:
        error:
          $ref: "#/components/schemas/ErrorObject"
      required:
      - error
      title: ErrorEnvelope
      type: object
...
    ReferenceListResponse:
      example:
        total: 0
        items:
        - metadata:
            key: ""
          code: code
          name: name
          id: id
        - metadata:
            key: ""
          code: code
          name: name
          id: id
      properties:
        items:
          items:
            $ref: "#/components/schemas/ReferenceItem"
          title: items
          type: array
        total:
          minimum: 0
          title: total
          type: integer
      required:
      - items
      - total
      title: ReferenceListResponse
      type: object
    Sku:
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        name: name
        description: description
        unit_of_measure: unit_of_measure
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        sku_code: sku_code
        status: active
      properties:
        id:
          title: id
          type: string
        sku_code:
          title: sku_code
          type: string
        name:
          title: name
          type: string
        description:
          title: description
          type: string
        unit_of_measure:
          title: unit_of_measure
          type: string
        status:
          $ref: "#/components/schemas/MasterStatus"
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      required:
      - created_at
      - description
      - id
      - name
      - sku_code
      - status
      - unit_of_measure
      - updated_at
      title: Sku
      type: object
    CreateSkuRequest:
      additionalProperties: false
      example:
        name: name
        description: description
        unit_of_measure: unit_of_measure
        sku_code: sku_code
        status: active
      properties:
        sku_code:
          title: sku_code
          type: string
        name:
          title: name
          type: string
        description:
          title: description
          type: string
        unit_of_measure:
          title: unit_of_measure
          type: string
        status:
          $ref: "#/components/schemas/MasterStatus"
      required:
      - description
      - name
      - sku_code
      - status
      - unit_of_measure
      title: CreateSkuRequest
      type: object
    UpdateSkuRequest:
      additionalProperties: false
      example:
        name: name
        description: description
        unit_of_measure: unit_of_measure
        sku_code: sku_code
        status: active
      minProperties: 1
      properties:
        sku_code:
          title: sku_code
          type: string
        name:
          title: name
          type: string
        description:
          title: description
          type: string
        unit_of_measure:
          title: unit_of_measure
          type: string
        status:
          $ref: "#/components/schemas/MasterStatus"
      title: UpdateSkuRequest
      type: object
    SkuListResponse:
      example:
        total: 0
        items:
        - updated_at: 2000-01-23T04:56:07.000+00:00
          name: name
          description: description
          unit_of_measure: unit_of_measure
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          sku_code: sku_code
          status: active
        - updated_at: 2000-01-23T04:56:07.000+00:00
          name: name
          description: description
          unit_of_measure: unit_of_measure
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          sku_code: sku_code
          status: active
      properties:
        items:
          items:
            $ref: "#/components/schemas/Sku"
          title: items
          type: array
        total:
          minimum: 0
          title: total
          type: integer
      required:
      - items
      - total
      title: SkuListResponse
      type: object
...
    Shipment:
      example:
        route_id: route_id
        eta: 2000-01-23T04:56:07.000+00:00
        updated_at: 2000-01-23T04:56:07.000+00:00
        sku_ids:
        - sku_ids
        - sku_ids
        destination_port_id: destination_port_id
        created_at: 2000-01-23T04:56:07.000+00:00
        shipment_code: shipment_code
        id: id
        origin_port_id: origin_port_id
        supplier_id: supplier_id
        status: planned
      properties:
        id:
          title: id
          type: string
        shipment_code:
          title: shipment_code
          type: string
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        origin_port_id:
          title: origin_port_id
          type: string
        destination_port_id:
          title: destination_port_id
          type: string
        route_id:
          title: route_id
          type: string
        supplier_id:
          title: supplier_id
          type: string
        sku_ids:
          items:
            type: string
          title: sku_ids
          type: array
        eta:
          format: date-time
          title: eta
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      required:
      - created_at
      - destination_port_id
      - eta
      - id
      - origin_port_id
      - route_id
      - shipment_code
      - sku_ids
      - status
      - supplier_id
      - updated_at
      title: Shipment
      type: object
    CreateShipmentRequest:
      additionalProperties: false
      example:
        route_id: route_id
        eta: 2000-01-23T04:56:07.000+00:00
        sku_ids:
        - sku_ids
        - sku_ids
        destination_port_id: destination_port_id
        shipment_code: shipment_code
        origin_port_id: origin_port_id
        supplier_id: supplier_id
        status: planned
      properties:
        shipment_code:
          title: shipment_code
          type: string
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        origin_port_id:
          title: origin_port_id
          type: string
        destination_port_id:
          title: destination_port_id
          type: string
        route_id:
          title: route_id
          type: string
        supplier_id:
          title: supplier_id
          type: string
        sku_ids:
          items:
            type: string
          title: sku_ids
          type: array
        eta:
          format: date-time
          title: eta
          type: string
      required:
      - destination_port_id
      - eta
      - origin_port_id
      - route_id
      - shipment_code
      - sku_ids
      - status
      - supplier_id
      title: CreateShipmentRequest
      type: object
    UpdateShipmentRequest:
      additionalProperties: false
      example:
        route_id: route_id
        eta: 2000-01-23T04:56:07.000+00:00
        sku_ids:
        - sku_ids
        - sku_ids
        destination_port_id: destination_port_id
        shipment_code: shipment_code
        origin_port_id: origin_port_id
        supplier_id: supplier_id
        status: planned
      minProperties: 1
      properties:
        shipment_code:
          title: shipment_code
          type: string
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        origin_port_id:
          title: origin_port_id
          type: string
        destination_port_id:
          title: destination_port_id
          type: string
        route_id:
          title: route_id
          type: string
        supplier_id:
          title: supplier_id
          type: string
        sku_ids:
          items:
            type: string
          title: sku_ids
          type: array
        eta:
          format: date-time
          title: eta
          type: string
      title: UpdateShipmentRequest
      type: object
    ShipmentListResponse:
      example:
        total: 0
        items:
        - route_id: route_id
          eta: 2000-01-23T04:56:07.000+00:00
          updated_at: 2000-01-23T04:56:07.000+00:00
          sku_ids:
          - sku_ids
          - sku_ids
          destination_port_id: destination_port_id
          created_at: 2000-01-23T04:56:07.000+00:00
          shipment_code: shipment_code
          id: id
          origin_port_id: origin_port_id
          supplier_id: supplier_id
          status: planned
        - route_id: route_id
          eta: 2000-01-23T04:56:07.000+00:00
          updated_at: 2000-01-23T04:56:07.000+00:00
          sku_ids:
          - sku_ids
          - sku_ids
          destination_port_id: destination_port_id
          created_at: 2000-01-23T04:56:07.000+00:00
          shipment_code: shipment_code
          id: id
          origin_port_id: origin_port_id
          supplier_id: supplier_id
          status: planned
      properties:
        items:
          items:
            $ref: "#/components/schemas/Shipment"
          title: items
          type: array
        total:
          minimum: 0
          title: total
          type: integer
      required:
      - items
      - total
      title: ShipmentListResponse
      type: object
...
    Incident:
      example:
        article_id: article_id
        risk_level: low
        updated_at: 2000-01-23T04:56:07.000+00:00
        risk_score: 8
        reasoning: reasoning
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        classification: risk_exposure
        overlap_tags:
        - port_overlap
        - port_overlap
        status: open
      properties:
        id:
          title: id
          type: string
        article_id:
          title: article_id
          type: string
        classification:
          $ref: "#/components/schemas/IncidentClassification"
        status:
          $ref: "#/components/schemas/IncidentStatus"
        reasoning:
          title: reasoning
          type: string
        overlap_tags:
          items:
            $ref: "#/components/schemas/OverlapTag"
          title: overlap_tags
          type: array
          uniqueItems: true
        risk_score:
          maximum: 100
          minimum: 0
          title: risk_score
          type: integer
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      required:
      - article_id
      - classification
      - created_at
      - id
      - overlap_tags
      - reasoning
      - risk_level
      - risk_score
      - status
      - updated_at
      title: Incident
      type: object
    IncidentListResponse:
      example:
        total: 0
        page: 1
        items:
        - article_id: article_id
          risk_level: low
          updated_at: 2000-01-23T04:56:07.000+00:00
          risk_score: 8
          reasoning: reasoning
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          classification: risk_exposure
          overlap_tags:
          - port_overlap
          - port_overlap
          status: open
        - article_id: article_id
          risk_level: low
          updated_at: 2000-01-23T04:56:07.000+00:00
          risk_score: 8
          reasoning: reasoning
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          classification: risk_exposure
          overlap_tags:
          - port_overlap
          - port_overlap
          status: open
        page_size: 60
      properties:
        items:
          items:
            $ref: "#/components/schemas/Incident"
          title: items
          type: array
        total:
          minimum: 0
          title: total
          type: integer
        page:
          minimum: 1
          title: page
          type: integer
        page_size:
          maximum: 100
          minimum: 1
          title: page_size
          type: integer
      required:
      - items
      - page
      - page_size
      - total
      title: IncidentListResponse
      type: object
...
    IncidentClassification:
      enum:
      - risk_exposure
      - active_disruption
      title: IncidentClassification
      type: string
    IncidentStatus:
      enum:
      - open
      - monitoring
      - awaiting_decision
      - resolved
      - dismissed
      title: IncidentStatus
      type: string
...
    RiskLevel:
      enum:
      - low
      - medium
      - high
      title: RiskLevel
      type: string
...
    ShipmentStatus:
      enum:
      - planned
      - in_transit
      - delayed
      - delivered
      - cancelled
      title: ShipmentStatus
      type: string
...
    OverlapTag:
      enum:
      - port_overlap
      - route_overlap
      - immediate_sku_threat
      title: OverlapTag
      type: string